fix bugs

+++ src/Message/Request.h
@@ -2,22 +2,22 @@
 #include <stdint.h>
 #include <arpa/inet.h>
 
-#define DISCOVER_REQUEST 0x80
-#define NOTCMD_REQUEST 0x00
+#define DISCOVER_REQUEST (uint8_t)0x80
+#define NOTCMD_REQUEST (uint8_t)0x00
 
-#define LIST_APP_REQUEST 0x09
-#define START_APP_REQUEST 0x0a
-#define STOP_APP_REQUEST 0x0b
+#define LIST_APP_REQUEST (uint8_t)0x09
+#define START_APP_REQUEST (uint8_t)0x0a
+#define STOP_APP_REQUEST (uint8_t)0x0b
 
-#define LIST_PROC_REQUEST 0x11
-#define KILL_PROC_REQUEST 0x12
+#define LIST_PROC_REQUEST (uint8_t)0x11
+#define KILL_PROC_REQUEST (uint8_t)0x12
 
-#define SCREENSHOT_REQUEST 0x19
+#define SCREENSHOT_REQUEST (uint8_t)0x19
 
-#define START_KEYLOG_REQUEST 0x21
-#define STOP_KEYLOG_REQUEST 0x22
+#define START_KEYLOG_REQUEST (uint8_t)0x21
+#define STOP_KEYLOG_REQUEST (uint8_t)0x22
 
-#define DIR_TREE_REQUEST 0x29
+#define DIR_TREE_REQUEST (uint8_t)0x29
 
 class Request {
     struct {
